minor fix and cleanup
